backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
python-dotenv
pytz
rest-framework-simplejwt
sqlparse
django-allauth
